# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/dashboard/summary": {
      "get": {
        "operationId": "DashboardDataController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DashboardSummaryDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Dashboard data is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Counts for the dashboard summary card",
        "tags": [
          "dashboard-data"
        ]
      }
    },
    "/api/dashboard/kpi": {
      "get": {
        "description": "Accepts sourceId, or the legacy metric alias the current frontend sends.",
        "operationId": "DashboardDataController_kpi",
        "parameters": [
          {
            "name": "sourceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sales-summary",
                "active-users",
                "traffic-series",
                "conversion-series",
                "recent-events",
                "incident-events"
              ]
            }
          },
          {
            "name": "metric",
            "required": false,
            "in": "query",
            "description": "Legacy alias resolved by the controller. Unknown values fall back to a default.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "month",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "quarter"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/KpiDataDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Dashboard data is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A single KPI value",
        "tags": [
          "dashboard-data"
        ]
      }
    },
    "/api/dashboard/chart": {
      "get": {
        "operationId": "DashboardDataController_chart",
        "parameters": [
          {
            "name": "sourceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sales-summary",
                "active-users",
                "traffic-series",
                "conversion-series",
                "recent-events",
                "incident-events"
              ]
            }
          },
          {
            "name": "metric",
            "required": false,
            "in": "query",
            "description": "Legacy alias resolved by the controller. Unknown values fall back to a default.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "month",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "quarter"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/SeriesDataDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Dashboard data is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A time series for a chart widget",
        "tags": [
          "dashboard-data"
        ]
      }
    },
    "/api/dashboard/table": {
      "get": {
        "operationId": "DashboardDataController_table",
        "parameters": [
          {
            "name": "sourceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sales-summary",
                "active-users",
                "traffic-series",
                "conversion-series",
                "recent-events",
                "incident-events"
              ]
            }
          },
          {
            "name": "metric",
            "required": false,
            "in": "query",
            "description": "Legacy alias resolved by the controller. Unknown values fall back to a default.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "month",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "quarter"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/TableDataDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Dashboard data is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rows for a table widget",
        "tags": [
          "dashboard-data"
        ]
      }
    },
    "/api/dashboards/{dashboardId}": {
      "get": {
        "description": "A dashboard the caller may not see answers 404, not 403 — a distinct 403 would confirm the id exists.",
        "operationId": "DashboardController_findOne",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DashboardResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such dashboard, or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read a dashboard definition",
        "tags": [
          "dashboards"
        ]
      },
      "put": {
        "description": "Requires the version last read. A mismatch answers 409 with the current version in details.",
        "operationId": "DashboardController_save",
        "parameters": [
          {
            "name": "dashboardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDashboardDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DashboardResponseDto"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The caller can see the dashboard but does not own it.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The dashboard changed since it was read.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "The definition failed schema validation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a dashboard definition",
        "tags": [
          "dashboards"
        ]
      }
    },
...
  },
  "info": {
    "title": "React Boilerplate API",
    "description": "Backend for the React boilerplate. Every response uses the shared envelope: `{ success: true, data }` on success and `{ success: false, error: { code, message, details? } }` on failure. HTTP status describes the transport outcome; `error.code` describes the domain outcome.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "rb_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rb_refresh"
      }
    },
    "schemas": {
...
      "DashboardSummaryDto": {
        "type": "object",
        "properties": {
          "activeUsers": {
            "type": "number"
          },
          "invitedUsers": {
            "type": "number"
          },
          "blockedUsers": {
            "type": "number"
          },
          "apiLatencyMs": {
            "type": "number"
          },
          "contractErrorRate": {
            "type": "number"
          }
        },
        "required": [
          "activeUsers",
          "invitedUsers",
          "blockedUsers",
          "apiLatencyMs",
          "contractErrorRate"
        ]
      },
      "KpiDataDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "kpi"
            ]
          },
          "label": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "trend": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "label"
        ]
      },
      "SeriesPointDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "value": {
            "type": "number"
          }
        },
        "required": [
          "label",
          "value"
        ]
      },
      "SeriesDataDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "series"
            ]
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeriesPointDto"
            }
          }
        },
        "required": [
          "kind",
          "points"
        ]
      },
      "TableColumnDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "event",
              "owner",
              "status"
            ]
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "label"
        ]
      },
      "TableRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "event": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "owner",
          "status"
        ]
      },
      "TableDataDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "table"
            ]
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TableColumnDto"
            }
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TableRowDto"
            }
          }
        },
        "required": [
          "kind",
          "columns",
          "rows"
        ]
      },
      "DashboardResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared"
            ]
          },
          "schemaVersion": {
            "type": "number",
            "description": "Schema version of the stored definition."
          },
          "version": {
            "type": "number",
            "description": "Optimistic lock. Send this back when writing."
          },
          "definition": {
            "type": "object",
            "description": "Validated dashboard definition."
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "ownerId",
          "visibility",
          "schemaVersion",
          "version",
          "definition",
          "updatedAt"
        ]
      },
      "SaveDashboardDefinitionDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "definition": {
            "type": "object"
          }
        },
        "required": [
          "expectedVersion",
          "definition"
        ]
      },
...
    }
  }
}